requirements.txt
psutil
pywin32
pystray
Pillow
pandas
keyboard>=0.13.5
